@@ -22,20 +22,20 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.39.2",
-    "@playwright/test": "^1.57.0",
-    "@types/express": "^4.17.25",
+    "@playwright/test": "^1.58.1",
+    "@types/express": "^5.0.6",
     "@types/jest": "^30.0.0",
     "@types/multer": "^2.0.0",
     "@types/node": "^22.x",
     "@types/supertest": "^6.0.3",
-    "eslint": "^8.57.1",
+    "eslint": "^9.39.2",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-n": "^17.23.1",
     "eslint-plugin-prettier": "^5.5.4",
     "eslint-plugin-promise": "^7.2.1",
     "eslint-plugin-simple-import-sort": "^12.1.1",
-    "globals": "^16.5.0",
+    "globals": "^17.3.0",
     "prettier": "^3.7.4",
     "supertest": "^7.1.4",
     "typescript-eslint": "^8.51.0"
@@ -43,7 +43,7 @@
   "dependencies": {
     "date-fns": "^4.1.0",
     "dotenv": "17.2.3",
-    "playwright": "^1.57.0",
+    "playwright": "^1.58.1",
     "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.3/xlsx-0.20.3.tgz"
   },
   "engines": {
